+++ src/private/IRremoteInt.h
@@ -49,11 +49,11 @@ typedef struct {
 } irparams_t;
 
 // ISR State-Machine : Receiver States
-#define IR_REC_STATE_IDLE      2
-#define IR_REC_STATE_MARK      3
-#define IR_REC_STATE_SPACE     4
-#define IR_REC_STATE_STOP      5
-#define IR_REC_STATE_OVERFLOW  6
+#define IR_REC_STATE_IDLE      0
+#define IR_REC_STATE_MARK      1
+#define IR_REC_STATE_SPACE     2
+#define IR_REC_STATE_STOP      3
+#define IR_REC_STATE_OVERFLOW  4
 
 /**
  * Allow all parts of the code access to the ISR data