+++ changelog.md
@@ -1,8 +1,12 @@
-## 2.5.0 2020/07
+## 2.5.1 2020/08
 - Added support for MagiQuest IR wands.
 - Corrected Samsung timing.
 - NEC repeat implemantation.
 - Formatting and changing TIMER_CONFIG_KHZ and TIMER_CONFIG_NORMAL macros to static functions.
+- Added IRAM_ATTR for ESP32 ISR.
+- Removed #define HAS_AVR_INTERRUPT_H.
+- Changed Receiver States. Now starting with 0.
+- Changed switch to if / else if in IRRemote,.cpp because of ESP32 compiler bug.
 
 ## 2.5.0 2020/06
 - corrected keywords.txt.